% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmr.R
\name{conectar}
\alias{conectar}
\title{Conecta a la base y mantiene abierta la sesión}
\usage{
conectar(
  usuario,
  password,
  servidor = "localhost",
  puerto = 27017,
  admindb = "admin"
)
}
\arguments{
\item{usuario}{nombre usuario}

\item{password}{password usuario}

\item{servidor}{ip servidor}

\item{puerto}{puerto conexion}

\item{admindb}{admin donde esta guarda la info de loge}
}
\description{
Conecta a la base y mantiene abierta la sesión
}
