% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmr.R
\name{conexion}
\alias{conexion}
\title{Conecta a la base y devuelve handler de la colección}
\usage{
conexion(
  coleccion,
  db,
  usuario,
  password,
  servidor = "localhost",
  puerto = 27017,
  admindb = "admin"
)
}
\arguments{
\item{coleccion}{nombre coleccion}

\item{db}{nombre db}

\item{usuario}{nombre usuario}

\item{password}{password usuario}

\item{servidor}{ip servidor}

\item{puerto}{puerto conexion}

\item{admindb}{admin donde esta guarda la info de loge}
}
\description{
Conecta a la base y devuelve handler de la colección
}
