% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmr.R
\name{contar}
\alias{contar}
\title{Cuenta noticias filtrando segun algunos parametros.}
\usage{
contar(
  desde,
  hasta,
  diarios = c(),
  categorias = c(),
  palabras_en_titulo = c(),
  palabras_en_texto = c()
)
}
\arguments{
\item{desde}{fecha desde}

\item{hasta}{fecha hasta}

\item{diarios}{diarios a buscar}

\item{categorias}{categorias a buscar}

\item{palabras_en_titulo}{palabras que deben aparecen en el titulo (pueden ser pedazos de palabras)}

\item{palabras_en_texto}{palabras que deben aparecen en el texto (pueden ser pedazos de palabras)}
}
\description{
Cuenta noticias filtrando segun algunos parametros.
}
