% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmr.R
\name{formatear}
\alias{formatear}
\title{Formatea la tabla según donde se quiera presentar}
\usage{
formatear(data, para = "blog", cantidad_decimales = 2)
}
\arguments{
\item{para}{tipo de formato segun donde se quiera publicar}

\item{cantidad_decimales}{cantidad de decimales a tomar. -1 para dejar la misma cantidad y que quede como numeric}

\item{tabla}{tabla a formatear}
}
\description{
Formatea la tabla según donde se quiera presentar
}
