% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmr.R
\name{foto}
\alias{foto}
\title{Devuelve una 'foto' que resumén el estado de datos en la base.}
\usage{
foto(diarios = c(), categorias = c(), desde = "00000000", hasta = "99999999")
}
\arguments{
\item{diarios}{diarios a filtrar. Si no se especifica, se toman todos.}

\item{categorias}{categorias a filtar. Si no se especifica, se toman todos.}

\item{desde}{fecha desde}

\item{hasta}{fecha hasta}
}
\description{
Devuelve una 'foto' que resumén el estado de datos en la base.
}
