% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmr.R
\name{foto_palabras}
\alias{foto_palabras}
\title{Devuelve una 'foto' de las palabras en un rango de tiempo.}
\usage{
foto_palabras(
  que,
  donde,
  diarios = c(),
  categorias = c(),
  desde = "00000000",
  hasta = "99999999",
  top = 2000,
  top_por_tendencia = 500,
  freq_min = 5,
  min_noticias = 1,
  por_categoria = T
)
}
\arguments{
\item{que}{qué buscar? 'personas', 'terminos' y/o 'verbos'}

\item{donde}{dónde buscar? 'textos' o 'titulos'}

\item{diarios}{diarios a filtrar. Si no se especifica, se toman todos.}

\item{categorias}{categorias a filtar. Si no se especifica, se toman todos.}

\item{desde}{fecha desde}

\item{hasta}{fecha hasta}

\item{top}{rankear las primeras N palabras}
}
\description{
Devuelve una 'foto' de las palabras en un rango de tiempo.
}
