% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmr.R
\name{frecuencias}
\alias{frecuencias}
\title{Busca 'personas', 'terminos' o 'verbos' en 'textos' o 'titulos' de la coleccion, filtrando segun algunos parametros.}
\usage{
frecuencias(
  conexion,
  que,
  donde,
  palabras,
  desde = "0",
  hasta = "99999999",
  diarios = c(),
  categorias = c(),
  freq_min = 0,
  freq_max = 99999
)
}
\arguments{
\item{conexion}{conexion a la coleccion 'frecuencias' ya instanciada}

\item{que}{qué buscar? 'personas', 'terminos' o 'verbos'}

\item{donde}{dónde buscar? 'textos' o 'titulos'}

\item{palabras}{palabras a buscar}

\item{desde}{fecha desde}

\item{hasta}{fecha hasta}

\item{diarios}{diarios a buscar}

\item{categorias}{categorias a buscar}

\item{freq_min}{freq minima de palabras a buscar}

\item{freq_max}{freq maxima de palabras a buscar}
}
\description{
Busca 'personas', 'terminos' o 'verbos' en 'textos' o 'titulos' de la coleccion, filtrando segun algunos parametros.
}
