% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdlmr.R
\name{gglollipop}
\alias{gglollipop}
\title{Dibuja los datos de una tabla tipo 'clave:valor' en un gr치fico estilo 'lollipop'.
El dataframe tiene que tener una columna 'clave' y otra 'valor'.}
\usage{
gglollipop(
  d,
  tamanio_texto = 15,
  colores = NULL,
  rango_y = NULL,
  separacion_lineas = 0.5,
  tamanio_puntos = 2,
  nombre_ejex = element_blank(),
  nombre_ejey = element_blank(),
  titulo = element_blank()
)
}
\arguments{
\item{d}{dataframe con tabla proveniente de la funcion 'tendencias'}

\item{titulo}{titulo del gr치fico}

\item{color}{color del gr치fico}
}
\description{
Dibuja los datos de una tabla tipo 'clave:valor' en un gr치fico estilo 'lollipop'.
El dataframe tiene que tener una columna 'clave' y otra 'valor'.
}
