% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmr.R
\name{tendencias}
\alias{tendencias}
\title{Busca 'personas', 'terminos' o 'verbos' en 'textos' o 'titulos' de la coleccion, filtrando segun algunos parametros.}
\usage{
tendencias(que, donde, desde, hasta, diarios = c(), categorias = c(), top = 10)
}
\arguments{
\item{que}{qué buscar? 'personas', 'terminos' o 'verbos'. Puede ser 1, 2 o los 3.}

\item{donde}{dónde buscar? 'textos' o 'titulos'}

\item{diarios}{diarios a buscar}

\item{categorias}{categorias a buscar}

\item{top}{traer las N palabras mas mencionadas}

\item{fecha}{fecha a medir tendencia}
}
\description{
Busca 'personas', 'terminos' o 'verbos' en 'textos' o 'titulos' de la coleccion, filtrando segun algunos parametros.
}
